Move block reduce to dedicated module
