+++ verde/__init__.py
@@ -8,8 +8,9 @@ from . import datasets
 from .base import BaseGridder
 from .coordinates import scatter_points, grid_coordinates, inside, \
     profile_coordinates, get_region
-from .grid_math import block_split, distance_mask, BlockReduce
-from .scipy_bridge import ScipyGridder
+from .grid_math import distance_mask
+from .blockreduce import block_split, BlockReduce
+from .scipygridder import ScipyGridder
 from .trend import Trend, trend_jacobian, VectorTrend
 from .chain import Chain
 from .spline import Spline, spline_jacobian
